prevent any path to be used
This commit ensures, that only the basename of a lyrics
file is saved to the user option. It will prevent a user
from using an arbitrary path and effectively using any
file on the file system as a lyrics file.

The selected lyrics file is also compared to the
available *.txt files in the lyrics folder in the
plugin and in the uploads folder.

This commit fixes CVE-2024-9224